requirements.txt
mysql-connector-python
jinja2
weasyprint
pdfkit
